refactorize endpoints
