+++ Gruntfile.js
@@ -15,9 +15,9 @@ module.exports = function (grunt) {
 			dist: ['dist']
 		},
 		copy: {
-			app: { expand: true, cwd: 'build/app', src: '**', dest: 'dist/' },
+			api: { expand: true, cwd: 'build/api', src: '**', dest: 'dist/' },
 			public: { expand: true, cwd: 'public', src: '**', dest: 'dist/public' },
-			// workspace: { expand: true, cwd: 'src/app/workspace', src: '**', dest: 'dist/workspace' },
+			// workspace: { expand: true, cwd: 'src/api/workspace', src: '**', dest: 'dist/workspace' },
 			license: { expand: true, src: './LICENSE', dest: 'dist/' }
 		}
 	})
@@ -36,6 +36,6 @@ module.exports = function (grunt) {
 	grunt.registerTask('build', ['clean:build', 'exec:swagger', 'exec:tsc'])
 	grunt.registerTask('lint', ['exec:lint'])
 	grunt.registerTask('doc', ['exec:typedoc'])
-	grunt.registerTask('dist', ['clean:dist', 'copy:app', 'copy:public', 'copy:license', 'create-package'])
+	grunt.registerTask('dist', ['clean:dist', 'copy:api', 'copy:public', 'copy:license', 'create-package'])
 	grunt.registerTask('default', [])
 }
