+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "lambdaorm-svc",
-	"version": "0.1.2",
+	"version": "0.1.3",
 	"description": "service of lambda orm",
 	"main": "./dist/lib/index",
 	"scripts": {
@@ -8,7 +8,7 @@
 		"build": "grunt build",
 		"doc": "grunt doc",
 		"dist": "grunt dist",
-		"start": "node build/app/index.js"
+		"start": "node build/api/index.js"
 	},
 	"nodemonConfig": {
 		"watch": [
@@ -40,7 +40,7 @@
 	"homepage": "https://github.com/FlavioLionelRita/lambdaorm-svc#readme",
 	"dependencies": {
 		"express": "^4.17.1",
-		"lambdaorm": "0.1.29-alpha",
+		"lambdaorm": "0.1.30-alpha",
 		"morgan": "^1.10.0",
 		"reflect-metadata": "^0.1.13",
 		"swagger-ui-express": "^4.1.6",
