+++ src/dev/lambdarmo-svc/docker-compose-win.yaml
@@ -25,7 +25,7 @@ services:
     depends_on:
       - db
     container_name: lambdaorm-svc
-    image: flaviorita/lambdaorm-svc:0.1.1
+    image: flaviorita/lambdaorm-svc:0.1.3
     restart: always
     environment:
       - HOST=http://localhost
